Finished up draft of splitter
There is one issue, that is the thresholds of
sklearn are a little different than ours.
I believe this is due to some optimization on their end
where they don't take the direct value of the splitting
feature, but do some further calculations to get more
information gain than just taking the splitting value.