Created the Predict class
Furthermore, I carried out some renaming of files, as it did not fit the standard of module names.
Files should not have the same name as the classes within.
Also moved the DepthTreeBuilder back in with the DecisionTree.
This is due to the fact that it causes circular imports if not,
how it was able to work previously is a mystery to myself.