Merge issues
Due to the changes made in filenames, the merge was unable to track changes between the "same"
files. Now I could have used git rebase, but in this case it was so far back, that it lead to other issues.
Ended up going through and moving the changes over manually with git diff.