Added deterministic seeding and fixed bugs
The multiprocessing didn't seem to work on my end. Had to do a lot of fixing. My guess would be
that the file has only been tested within the same document in a if __name__ == "__main__" statement,
as it didn't work beyond that. It should be fixed now, although in a janky way atm.