Fix to setup.py editing mode
I accidentally misunderstood the Extension type, and added the python
modules to it. This meant that the python was actually compiled into C
when used. I have removed this functionality and kept it as python. This
allows for the use of editable mode with pip significantly speeding up
the development of python files.